Rejoin member
Category 1.) alumnis who followed the proper procedures to leave "in good standing" don't require mod checks, bootcamps etc. Thus they are in the pervue of the Member Manager (who doesn't know do mod checks or bootcamps)

Category 2.) is an ex-member but they were discharged to due failing to respond to multiple messages about their low attendance, these members may be able to join back but they need a.) the clearing of the Interviewing Manager and b.) are expected in most cases to start as Recruits again since they didn't leave in good standing. Alternatively Recruits who didn't finish their Recruitment phase would also be in this category. Thus in the pervue of Interviewing Branch.

Category 3.) banned members are only rare cases of repeated bad conduct, thus members are able to appeal their ban but it would be subject to collective SSgt decision since they handed out the ban, and is thus in the pervue of SSgts.